SREP-1865 - Update synthetic monitoring when API ingress changes
