+++ go.mod
@@ -6,6 +6,7 @@ toolchain go1.23.8
 
 require (
 	github.com/go-logr/logr v1.4.2
+	github.com/golang/mock v1.6.0
 	github.com/google/gofuzz v1.2.0
 	github.com/hashicorp/go-version v1.6.0
 	github.com/onsi/ginkgo v1.16.5
